+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0 - 2025-03-19]
+## [1.0.0 - 2025-03-24]
 ### Added
 - Initial commit
 - Init project
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Spring 3.1.0
 - Rest service "variable"
 - Rest service "variables"
+- Log reachable or unreachable ping
 
 ### Fixed
 - Filter external links of import map